+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #0.0.3
 - Adding pubspec snippet
-
+- Adding `pub run` for the current document
 
 #0.0.2
 - Adding `pub get` command